+++ zed_components/src/zed_camera/include/sl_types.hpp
@@ -47,6 +47,7 @@
 #include <std_msgs/msg/bool.hpp>
 #include <std_srvs/srv/set_bool.hpp>
 #include <std_srvs/srv/trigger.hpp>
+#include <cob_srvs/srv/set_int.hpp>
 #include <stereo_msgs/msg/disparity_image.hpp>
 #include <string>
 #include <vector>
@@ -177,11 +178,13 @@ typedef rclcpp::Service<zed_msgs::srv::SetPose>::SharedPtr setPoseSrvPtr;
 typedef rclcpp::Service<std_srvs::srv::SetBool>::SharedPtr enableObjDetPtr;
 typedef rclcpp::Service<std_srvs::srv::SetBool>::SharedPtr enableBodyTrkPtr;
 typedef rclcpp::Service<std_srvs::srv::SetBool>::SharedPtr enableMappingPtr;
+
 typedef rclcpp::Service<zed_msgs::srv::StartSvoRec>::SharedPtr
   startSvoRecSrvPtr;
 typedef rclcpp::Service<zed_msgs::srv::SetROI>::SharedPtr setRoiSrvPtr;
 typedef rclcpp::Service<std_srvs::srv::Trigger>::SharedPtr stopSvoRecSrvPtr;
 typedef rclcpp::Service<std_srvs::srv::Trigger>::SharedPtr pauseSvoSrvPtr;
+typedef rclcpp::Service<cob_srvs::srv::SetInt>::SharedPtr setSvoFramePtr;
 typedef rclcpp::Service<std_srvs::srv::Trigger>::SharedPtr resetRoiSrvPtr;
 typedef rclcpp::Service<robot_localization::srv::ToLL>::SharedPtr toLLSrvPtr;
 typedef rclcpp::Service<robot_localization::srv::FromLL>::SharedPtr fromLLSrvPtr;
@@ -210,6 +213,7 @@ std::string toString(const PubRes & res)
 
 typedef enum
 {
+  PUB,     //!< Same resolution as Color and Depth Map. [Old behavior for compatibility]
   FULL,    //!< Full resolution. Not recommended because slow processing and high bandwidth requirements
   COMPACT,  //!< Standard resolution. Optimizes processing and bandwidth
   REDUCED   //!< Half resolution. Low processing and bandwidth requirements
@@ -217,6 +221,8 @@ typedef enum
 std::string toString(const PcRes & res)
 {
   switch (res) {
+    case PUB:
+      return "PUB";
     case FULL:
       return "FULL";
     case COMPACT: